+++ CHANGELOG.md
@@ -3,6 +3,7 @@ _Note: Numbers like (\#123) point to closed Pull Requests on the fractal-web rep
 # Unreleased
 
 - Support starring datasets (\#1100).
+- Use new combined latest-job & status endpoint (\#1103)
 
 # 1.27.6